+++ src/Components/Sign/SignUpContainer/index.scss
@@ -1,0 +1,74 @@
+/*
+  Sign Up Container Styling
+*/
+
+.signUpContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px 0;
+  background-color: $white;
+  max-width: 480px;
+  height: auto;
+  border-radius: 25px;
+
+  &__form {
+    @include content-center;
+    width: 80%;
+
+    &-linked {
+      display: flex;
+      justify-content: space-around;
+      padding: 10px 0 20px 0;
+      width: 100%;
+    }
+
+    &-division {
+      display: flex;
+      width: 100%;
+
+      & hr {
+        width: 100%;
+      }
+
+      & span {
+        font-size: $guidance;
+        padding: 5px 10px 0 10px;
+      }
+    }
+
+    &-policy {
+      font-size: $description;
+      text-align: center;
+
+      & hr {
+        width: 60%;
+        margin-top: 0;
+      }
+    }
+  }
+
+  &__footer {
+    @include content-center;
+    margin-top: 20px;
+
+    &-signinlink {
+      font-size: $guidance;
+      padding-top: 10px;
+
+      & a {
+        font-weight: 600;
+        color: $main-color;
+        border-bottom: 1px solid $main-color;
+        margin-left: 8px;
+
+        &:hover {
+          font-weight: 700;
+          color: $main-color;
+          border: 0;
+        }
+      }
+    }
+  }
+}
