+++ src/index.scss
@@ -13,6 +13,7 @@
 @import 'Components/Thumbnail';
 @import 'Components/Sign/SignInForm';
 @import 'Components/Sign/SignUpForm';
+@import 'Components/Sign/SignUpContainer';
 @import 'Components/Upload/UploadPost';
 @import 'Components/Search/';
 @import 'Components/Post/Post';
